New directory to include "workflows"
Workflows document how the CDM is queried for specific data.
These worklflows could be adopted in Web/UI front-ends to
provide your users with the ability to generate reports.
For example, one of the most common uses would be to
generate a result (a value for the primary_metric) for a
benchmark iteration (a benchmark run in a specific way)
from a benchmark run (one user's run of a benchmark harness
like pbench or browbeat).